project_expense_tracking/requirement.txt
streamlit==1.48.1
pandas==2.2.3
fastapi==0.116.1
pytest==8.4.1
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.3
uvicorn==0.35.0
mysql-connector-python==9.4.0